# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookshelf API",
        "description": "API documentation for Bookshelf application",
        "contact": {
            "email": "support@test.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books/{bookId}/chapters": {
            "get": {
                "tags": [
                    "Chapter"
                ],
                "summary": "Get all chapters of a book",
                "operationId": "0d4ec9c723237cc1b44b7f25ee7919d0",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chapter"
                ],
                "summary": "Create a new chapter in a book",
                "operationId": "46fd1214cc2888ecdff41589ae0fb9ea",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "chapter_number": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chapter created successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/books/{bookId}/chapters/{chapterId}": {
            "get": {
                "tags": [
                    "Chapter"
                ],
                "summary": "Get a specific chapter of a book",
                "operationId": "3509ca4c9a04fabf0c6b5ccb88f56163",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Chapter"
                ],
                "summary": "Update a specific chapter of a book",
                "operationId": "90d34ac8d0ff93b3d2499308b3f14527",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "chapter_number": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chapter updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chapter"
                ],
                "summary": "Delete a specific chapter of a book",
                "operationId": "60769be01a163dd78f02faa01e64581a",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chapter deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/chapters/{chapterId}/full-content": {
            "get": {
                "tags": [
                    "Chapter"
                ],
                "summary": "Get full content of a chapter",
                "operationId": "2831ba1b16b25684fc2881a2c03cf84f",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    }
}